+++ src/tests.rs
@@ -1,4 +1,4 @@
-use crate::{Discard, Logger};
+use crate::{AsFmtSerializer, Discard, Drain, Logger, Never, OwnedKVList, Record, KV};
 
 // Separate module to test lack of imports
 mod no_imports {
@@ -19,6 +19,74 @@ mod no_imports {
 mod std_only {
     use super::super::*;
     use std;
+
+    #[derive(Clone)]
+    struct CheckError;
+
+    impl Drain for CheckError {
+        type Ok = ();
+        type Err = Never;
+        fn log(
+            &self,
+            record: &Record,
+            values: &OwnedKVList,
+        ) -> std::result::Result<Self::Ok, Self::Err> {
+            struct ErrorSerializer(String);
+
+            impl Serializer for ErrorSerializer {
+                fn emit_arguments(
+                    &mut self,
+                    key: Key,
+                    val: &fmt::Arguments,
+                ) -> Result {
+                    use core::fmt::Write;
+
+                    match key {
+                        "error" => self.0.write_fmt(*val).unwrap(),
+                        _ => {
+                            self.0.write_str(&key).unwrap();
+                            self.0.write_str(": ").unwrap();
+                            self.0.write_fmt(*val).unwrap();
+                            self.0.write_str("; ").unwrap();
+                        }
+                    }
+                    Ok(())
+                }
+            }
+
+            let mut serializer = ErrorSerializer(String::new());
+            values.serialize(record, &mut serializer).unwrap();
+            assert_eq!(serializer.0, format!("{}", record.msg()));
+            Ok(())
+        }
+    }
+
+    #[derive(Debug)]
+    struct TestError<E = std::string::ParseError>(&'static str, Option<E>);
+
+    impl TestError {
+        fn new(message: &'static str) -> Self {
+            TestError(message, None)
+        }
+    }
+
+    impl<E> fmt::Display for TestError<E> {
+        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+            write!(f, "{}", self.0)
+        }
+    }
+
+    impl<E: std::error::Error + 'static> std::error::Error for TestError<E> {
+        #[allow(deprecated)]
+        fn cause(&self) -> Option<&std::error::Error> {
+            self.1.as_ref().map(|error| error as _)
+        }
+
+        #[allow(deprecated)]
+        fn description(&self) -> &str {
+            "test error"
+        }
+    }
 
     #[test]
     fn logger_fmt_debug_sanity() {
@@ -58,6 +126,59 @@ mod std_only {
         info!(log, "(c, b2, b1, a)");
         let log = Logger::root(log, o!("d1" => "dd", "d2" => "dd"));
         info!(log, "(d2, d1, c, b2, b1, a)");
+    }
+
+    #[test]
+    fn error_fmt_no_source() {
+        let logger =
+            Logger::root(CheckError, o!("error" => #TestError::new("foo")));
+        info!(logger, "foo");
+        slog_info!(logger, "foo");
+    }
+
+    #[test]
+    fn error_fmt_no_source_not_last() {
+        let logger = Logger::root(
+            CheckError,
+            o!("error" => #TestError::new("foo"), "not-error" => "not-error"),
+        );
+        info!(logger, "not-error: not-error; foo");
+        slog_info!(logger, "not-error: not-error; foo");
+    }
+
+    #[test]
+    fn error_fmt_no_source_last() {
+        let logger = Logger::root(
+            CheckError,
+            o!("not-error" => "not-error", "error" => #TestError::new("foo")),
+        );
+        info!(logger, "foonot-error: not-error; ");
+        slog_info!(logger, "foonot-error: not-error; ");
+    }
+    #[test]
+    fn error_fmt_single_source() {
+        let logger = Logger::root(
+            CheckError,
+            o!("error" => #TestError("foo", Some(TestError::new("bar")))),
+        );
+        info!(logger, "foo: bar");
+    }
+
+    #[test]
+    fn error_fmt_two_sources() {
+        let logger = Logger::root(
+            CheckError,
+            o!("error" => #TestError("foo", Some(TestError("bar", Some(TestError::new("baz")))))),
+        );
+        info!(logger, "foo: bar: baz");
+    }
+
+    #[test]
+    fn ioerror_impls_value() {
+        let logger = Logger::root(Discard, o!());
+        info!(logger, "not found"; "error" => std::io::Error::from(std::io::ErrorKind::NotFound));
+        // compiles?
+        info!(logger, "not found"; "error" => #std::io::Error::from(std::io::ErrorKind::NotFound));
     }
 }
 
@@ -206,6 +327,67 @@ fn expressions_fmt() {
     info!(log, "message"; "f" => %f, "d" => ?d);
 }
 
+#[cfg(feature = "std")]
+#[test]
+fn display_and_alternate_display() {
+    use core::cell::Cell;
+    use core::fmt;
+
+    struct Example;
+
+    impl fmt::Display for Example {
+        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+            if f.alternate() {
+                f.write_str("alternate")
+            } else {
+                f.write_str("normal")
+            }
+        }
+    }
+
+    #[derive(Clone, Default)]
+    struct CheckExample;
+
+    impl Drain for CheckExample {
+        type Ok = ();
+        type Err = Never;
+
+        fn log(
+            &self,
+            record: &Record,
+            values: &OwnedKVList,
+        ) -> Result<(), Never> {
+            let mut checked_n = false;
+            let mut checked_a = false;
+            {
+                let mut serializer = AsFmtSerializer(|key, fmt_args| {
+                    if key == "n" {
+                        assert_eq!(format!("{}", fmt_args), "normal");
+                        checked_n = true;
+                    } else if key == "a" {
+                        assert_eq!(format!("{}", fmt_args), "alternate");
+                        checked_a = true;
+                    } else {
+                        panic!("Unexpected key: {}", key);
+                    }
+                    Ok(())
+                });
+
+                record.kv.serialize(record, &mut serializer).unwrap();
+            }
+
+            assert!(checked_n, "Expected the normal formatter to be used");
+            assert!(checked_a, "Expected the alternate formatter to be used");
+
+            Ok(())
+        }
+    }
+
+    let log = Logger::root(CheckExample, o!());
+
+    info!(log, ""; "n" => %Example, "a" => #%Example);
+}
+
 #[test]
 fn makers() {
     use crate::*;
@@ -223,7 +405,6 @@ fn makers() {
 #[test]
 fn simple_logger_erased() {
     use crate::*;
-
     fn takes_arced_drain(_l: Logger) {}
 
     let drain = Discard.filter_level(Level::Warning).map(Fuse);
@@ -259,4 +440,22 @@ fn logger_by_ref() {
     let f = "f";
     let d = (1, 2);
     info!(&log, "message"; "f" => %f, "d" => ?d);
+}
+
+#[test]
+fn test_never_type_clone() {
+    // We just want to make sure that this compiles
+    fn _do_not_run() {
+        let x: Never = panic!("Can't actually construct a Never type here!");
+        let y = x.clone();
+    }
+    // Always pass if we compiled
+}
+
+#[cfg(feature = "std")]
+#[test]
+fn can_hash_keys() {
+    use std::collections::HashSet;
+    use Key;
+    let tab: HashSet<Key> = ["foo"].iter().map(|&k| k.into()).collect();
 }